requirements.txt
dj-database-url==0.4.1
Django==1.9.7
gunicorn==19.6.0
psycopg2==2.6.2
whitenoise==2.0.6
requests==2.9.1
praw==4.4.0
python-twitter==3.2.1
branchio==0.1.1
